[Fix/#30] ShieldConfig 오류 해결
