+++ HMH_iOS/ShieldConfigureExtension/ShieldConfigureExtension.entitlements
@@ -1,12 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
-<dict>
-	<key>com.apple.developer.family-controls</key>
-	<true/>
-	<key>com.apple.security.application-groups</key>
-	<array>
-		<string>group.HMH</string>
-	</array>
-</dict>
+    <dict>
+        <key>com.apple.developer.family-controls</key>
+        <true/>
+    </dict>
 </plist>
